requirements.txt
pymongo>=4.0.0
google-generativeai>=0.3.0
nltk>=3.8.1
rouge>=1.0.1
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
reportlab>=3.6.0
chromadb>=0.4.0